Revision of the coordinate systems section
I tried to cut this down a bit. It's kind of boring to read because of
all the definitions and it's not the focus of this letter. The letter
should be straight to the point and only define things that are need for
our arguments.
I removed the conversion from geocentric to topocentric (eq 2) because
it isn't used anywhere else.
If we liberate some space, we can make room for more figures.