Use the vector symbol to make things clearer
It's very hard to see the bold gamma. I added the `\vec` command to draw
the little vector arrow to make things cleared. Kept the bold, like what
was used for the unit vectors.
I made the `\versor` and `\vector` commands to make it easier to control
the formatting for all the text.