Changes to definition of anomaly and disturbance
Simplified some of the text and added some missing italics.
Removed the P subscript from the gravity anomaly because it is not
defined on the point P (which is where the disturbance is defined).
The anomaly can't be defined on a point in 3D space, as we comment right
below that. So it's best to just say Delta g or Delta g(lambda, phi).

Added a part about not being able to upward continue Delta g because
it's not harmonic. This was implied by a sentence below saying that we
can upcontinue the disturbance bc it is harmonic.

Rephrased the part about authors doing upward continuation of the
anomaly. That might offend some people. Instead, I'm using the passive
voice to avoid putting blame on anyone.