Changes to diff between disturbance and anomaly
When defining the free-air anomaly, the reduction is applied to g, not
gamma. It's reducing g from P to Q', not raising gamma.

I really don't like the term "geoidal undulation" (it has nothing to do
with waves) or, even worse, "geoid anomaly". N is a distance so "geoidal
height" is better and is what is used in the Physical Geodesy book.