Reworked the part about Bouguer and its meaning
This part was a bit disjointed and it wasn't address the important
question about the meaning of the corrections. In applied geophysics,
Bouguer is more important that free-air and more attention should be
given to it.
I put in the equation for a Bouguer disturbance to show that we should
use geometric heights for topographic correction as well.
Add a recommendation to use just use h instead of H to the end of the
section.