Remove the mentions of the purpose of reductions
Modifications made after our live discussion.

There is no reason to get into that in this paper. We're already
polemical enough.
Changed the title back to a question.
Part about free-air is a discussion of it being an approximation to the
disturbance so moved it back to the previous section.
Bouguer is now a section on the topographic masses.